setup permission for routes

+++ backend/db/repository/blog.py
@@ -25,7 +25,9 @@ def list_blogs(db: Session):
 def update_blog(id: int, blog: UpdateBlog, author_id: int, db: Session):
     blog_in_db = db.query(Blog).filter(Blog.id == id).first()
     if not blog_in_db:
-        return
+        return {"error": f"Blog with id {id} does not exist"}
+    if not blog_in_db.author_id == author_id:
+        return {"error": "Only the author can modify the blog"}
     blog_in_db.title = blog.title
     blog_in_db.content = blog.content
     db.add(blog_in_db)
@@ -37,6 +39,8 @@ def delete_blog(id: int, author_id: int, db: Session):
     blog_in_db = db.query(Blog).filter(Blog.id == id)
     if not blog_in_db.first():
         return {"error": f"Could not find blog with id {id}"}
+    if not blog_in_db.first().author_id == author_id:
+        return {"error": "Only the author can delete a blog"}
     blog_in_db.delete()
     db.commit()
     return {"msg": f"deleted blog with id {id}"}
